#321 bootstrap module separation, manual conversions also available. TODO: how may we decopypaste typeclasses?
