+++ distage/distage-config/src/main/scala/com/github/pshirshov/izumi/distage/config/ConfigModule.scala
@@ -2,12 +2,12 @@ package com.github.pshirshov.izumi.distage.config
 
 import com.github.pshirshov.izumi.distage.config.codec.{RuntimeConfigReader, RuntimeConfigReaderDefaultImpl}
 import com.github.pshirshov.izumi.distage.config.model.AppConfig
-import com.github.pshirshov.izumi.distage.model.definition.ModuleDef
+import com.github.pshirshov.izumi.distage.model.definition.BootstrapModuleDef
 import com.github.pshirshov.izumi.distage.model.planning.PlanningHook
 
 case class ConfigInjectorConfig(enableScalars: Boolean = false)
 
-class ConfigModule(config: AppConfig, configInjectorConfig: ConfigInjectorConfig = ConfigInjectorConfig()) extends ModuleDef {
+class ConfigModule(config: AppConfig, configInjectorConfig: ConfigInjectorConfig = ConfigInjectorConfig()) extends BootstrapModuleDef {
   make[ConfigInjectorConfig].from(configInjectorConfig)
   make[AppConfig].from(config)
   make[RuntimeConfigReader].from[RuntimeConfigReaderDefaultImpl]
