+++ distage/distage-model/src/main/scala/com/github/pshirshov/izumi/distage/model/definition/ModuleApi.scala
@@ -1,0 +1,7 @@
+package com.github.pshirshov.izumi.distage.model.definition
+
+trait ModuleApi[T <: ModuleBase] {
+  def empty: T
+
+  def simple(bindings: Set[Binding]): T
+}
