--- distage/distage-model/src/main/scala/com/github/pshirshov/izumi/distage/model/definition/SimpleModuleDef.scala
@@ -1,7 +1,0 @@
-package com.github.pshirshov.izumi.distage.model.definition
-
-final case class SimpleModuleDef(bindings: Set[Binding] = Set.empty[Binding]) extends ModuleBase
-
-object SimpleModuleDef {
-  def empty: SimpleModuleDef = SimpleModuleDef(Set.empty[Binding])
-}
