+++ distage/distage-plugins/src/main/scala/com/github/pshirshov/izumi/distage/plugins/merge/PluginMergeStrategy.scala
@@ -4,5 +4,5 @@ import com.github.pshirshov.izumi.distage.plugins.{LoadedPlugins, PluginBase}
 
 
 trait PluginMergeStrategy[T <: LoadedPlugins] {
-  def merge[D <: PluginBase](defs: Seq[D]): T
+  def merge(defs: Seq[PluginBase]): T
 }
