+++ distage/distage-static/src/test/scala/com/github/pshirshov/izumi/distage/staticinjector/MkInjector.scala
@@ -1,10 +1,11 @@
 package com.github.pshirshov.izumi.distage.staticinjector
 
-import distage.{Injector, ModuleBase}
+
+import distage.{Injector, BootstrapModule}
 
 trait MkInjector {
 
-  def mkInjector(overrides: ModuleBase*): Injector = Injector.noReflection(overrides: _*)
+  def mkInjector(overrides: BootstrapModule*): Injector = Injector.noReflection(overrides: _*)
 
   def mkInjectorWithProxy(): Injector = Injector()
